Added new class for commented methods, updated BusTicketService class and jar file.
